deprecate src/contrib/torch/base64.h and use src/support/base64.h
